Move ReallyEqualMixin and s31e back to common_util
s31e = skip_if_cannot_represent_filename
